[networking] Move NetworkManager content to a new plugin
As NetworkManager is non-essential to configuring a network on most
distributions, and it has its own concepts of how to handle connections
different than the traditional means of ifconfig and ip, move anything
NetworkManager-related to a separate plugin.

Signed off by: Robb Manes <robb.manes@gmail.com>
